Refactor transaction inherent methods using macros and optimize PropertyMap display
