+++ src/db/tests.rs
@@ -1,5 +1,5 @@
 use super::*;
-use crate::api::transaction::{ReadOps, WriteOps};
+
 use crate::core::GLOBAL_INTERNER;
 use crate::core::error::{Error, Result};
 use crate::core::id::NodeId;
